fix: Add new mutation to get zoom meeting url
